feat: work on the tweets
Ditched the tweetsController as it fits more to be a part the usersController

in general, I'll search for users by the username not _id as it is more convenient and Twitter seems to be doing this. At least from the client prospective, that is.

deleting/patching tweets will be by _id, though.

Lastly, just some basic implementation for some query parameters like sort, limit, page etc.